requirements.txt
ephem
pandas
pillow
python-pptx
requests_pkcs12
staticmap
streamlit
bs4
markdown